fixes #20 by conditionally installing xmlrpc at install time. (#22)
* fixes #20 by conditionally installing xmlrpc at install time.

A bit of background:
in ruby version 2.4 they removed xmlrpc from std lib. So we need to conditionally install it depending on what version of ruby is being used on the target system (not the build system).

It seeems tempting (and wrong) to say just put this in the `.gemspec` but this is evaluated at **build** time and we want to check this at **install** time.

To my limited knowledge the only way to do this is to create a ruby extension to check and install at install time. I have commented in the `.gemspec` about this so that its less "silent" when someone is looking through it. I have also included 2 dockerfiles that are meant to be used to test build and install of gem on ruby 2.2-2.4 and have appropriate information in README on how to use.

* adding ruby 2.4 to travis